Cleanup first a bit
